+++ packages/react-dom/npm/react-dom.react-server.js
@@ -1,0 +1,7 @@
+'use strict';
+
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./cjs/react-dom.react-server.production.min.js');
+} else {
+  module.exports = require('./cjs/react-dom.react-server.development.js');
+}
