+++ packages/react-server-dom-turbopack/src/__tests__/ReactFlightTurbopackDOMForm-test.js
@@ -33,7 +33,7 @@ describe('ReactFlightDOMForm', () => {
   beforeEach(() => {
     jest.resetModules();
     // Simulate the condition resolution
-    jest.mock('react', () => require('react/react.shared-subset'));
+    jest.mock('react', () => require('react/react.react-server'));
     jest.mock('react-server-dom-turbopack/server', () =>
       require('react-server-dom-turbopack/server.edge'),
     );
