+++ packages/react-server-dom-webpack/src/__tests__/ReactFlightDOMBrowser-test.js
@@ -36,7 +36,7 @@ describe('ReactFlightDOMBrowser', () => {
     jest.resetModules();
 
     // Simulate the condition resolution
-    jest.mock('react', () => require('react/react.shared-subset'));
+    jest.mock('react', () => require('react/react.react-server'));
     jest.mock('react-server-dom-webpack/server', () =>
       require('react-server-dom-webpack/server.browser'),
     );
