+++ packages/react-server-dom-webpack/src/__tests__/ReactFlightDOMForm-test.js
@@ -38,7 +38,7 @@ describe('ReactFlightDOMForm', () => {
   beforeEach(() => {
     jest.resetModules();
     // Simulate the condition resolution
-    jest.mock('react', () => require('react/react.shared-subset'));
+    jest.mock('react', () => require('react/react.react-server'));
     jest.mock('react-server-dom-webpack/server', () =>
       require('react-server-dom-webpack/server.edge'),
     );
