+++ packages/react-server-dom-webpack/src/__tests__/ReactFlightDOMNode-test.js
@@ -29,7 +29,7 @@ describe('ReactFlightDOMNode', () => {
     jest.resetModules();
 
     // Simulate the condition resolution
-    jest.mock('react', () => require('react/react.shared-subset'));
+    jest.mock('react', () => require('react/react.react-server'));
     jest.mock('react-server-dom-webpack/server', () =>
       require('react-server-dom-webpack/server.node'),
     );
