+++ packages/react-server-dom-webpack/src/__tests__/ReactFlightDOMReply-test.js
@@ -24,7 +24,7 @@ describe('ReactFlightDOMReply', () => {
   beforeEach(() => {
     jest.resetModules();
     // Simulate the condition resolution
-    jest.mock('react', () => require('react/react.shared-subset'));
+    jest.mock('react', () => require('react/react.react-server'));
     jest.mock('react-server-dom-webpack/server', () =>
       require('react-server-dom-webpack/server.browser'),
     );
