+++ packages/react/react.react-server.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
+ * Copyright (c) Meta Platforms, Inc. and its affiliates.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -7,5 +7,4 @@
  * @flow
  */
 
-export * from './ReactSharedSubset';
-export {jsx, jsxs, jsxDEV} from './jsx/ReactJSX';
+export * from './src/ReactServer';
