+++ packages/react/src/ReactServerFB.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) Meta Platforms, Inc. and its affiliates.
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -7,4 +7,5 @@
  * @flow
  */
 
-export * from './src/ReactSharedSubset';
+export * from './ReactServer';
+export {jsx, jsxs, jsxDEV} from './jsx/ReactJSX';
